recommendation-service/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==2.2.5
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.33
protobuf==3.20.3
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
